feat: introduce admin constants and refactor hooks for improved organization
- Added new constants for product and coupon forms, validation limits, and admin tab information in a dedicated admin.ts file.
- Refactored useCouponForm and useProductForm hooks to utilize the new constants for initial form states.
- Updated useProducts and useNotification hooks to implement a unique ID generation utility for better ID management.
- Cleaned up imports in AdminPage to reference the new constants file, enhancing code clarity and maintainability.
