+++ src/basic/hooks/useNotification.ts
@@ -1,4 +1,5 @@
 import { useState, useCallback } from "react";
+import { useUniqueId } from "../utils/hooks/useUniqueId";
 
 interface Notification {
   id: string;
@@ -8,24 +9,21 @@ interface Notification {
 
 const TIMEOUT = 3000 as const;
 
-// 고유한 ID 생성 함수
-let notificationIdCounter = 0;
-const generateNotificationId = () => {
-  notificationIdCounter += 1;
-  return `${Date.now()}_${notificationIdCounter}`;
-};
-
 export const useNotification = () => {
   const [notifications, setNotifications] = useState<Notification[]>([]);
+  const generateId = useUniqueId();
 
-  const addNotification = useCallback((message: string, type: "error" | "success" | "warning" = "success") => {
-    const id = generateNotificationId();
-    setNotifications((prev) => [...prev, { id, message, type }]);
+  const addNotification = useCallback(
+    (message: string, type: "error" | "success" | "warning" = "success") => {
+      const id = generateId();
+      setNotifications((prev) => [...prev, { id: `notification_${id}`, message, type }]);
 
-    setTimeout(() => {
-      setNotifications((prev) => prev.filter((n) => n.id !== id));
-    }, TIMEOUT);
-  }, []);
+      setTimeout(() => {
+        setNotifications((prev) => prev.filter((n) => n.id !== id));
+      }, TIMEOUT);
+    },
+    [generateId]
+  );
 
   const removeNotification = useCallback((id: string) => {
     setNotifications((prev) => prev.filter((n) => n.id !== id));
