+++ src/basic/utils/hooks/useUniqueId.ts
@@ -1,0 +1,9 @@
+import { useId } from "react";
+
+export const useUniqueId = () => {
+  const baseId = useId();
+
+  return () => {
+    return `${baseId}`;
+  };
+};
